added HTML5 load error tests to shared tests
