+++ spec/mediaelement-webaudio.spec.js
@@ -1,8 +1,13 @@
 /* eslint-env jasmine */
 
-import { sharedTests } from './mediaelement-shared';
+import { sharedErrorTests, sharedTests } from './mediaelement-shared';
 
 /** @test {WaveSurfer} */
 describe('WaveSurfer/MediaElementWebAudio:', function() {
     sharedTests('MediaElementWebAudio');
+});
+
+/** @test {WaveSurfer} */
+describe('WaveSurfer/MediaElementWebAudio/errors:', function() {
+    sharedErrorTests('MediaElementWebAudio');
 });
