+++ spec/mediaelement.spec.js
@@ -1,6 +1,6 @@
 /* eslint-env jasmine */
 
-import { sharedTests } from './mediaelement-shared';
+import { sharedErrorTests, sharedTests } from './mediaelement-shared';
 import TestHelpers from './test-helpers';
 
 /** @test {WaveSurfer} */
@@ -10,28 +10,5 @@ describe('WaveSurfer/MediaElement:', function() {
 
 /** @test {WaveSurfer} */
 describe('WaveSurfer/MediaElement/errors:', function() {
-    var element;
-
-    beforeEach(function() {
-        element = TestHelpers.createElement('test');
-    });
-
-    afterEach(function() {
-        TestHelpers.removeElement(element);
-    });
-
-    /**
-     * @test {WaveSurfer}
-     */
-    it('throw when load is called with invalid HTMLMediaElement', function() {
-        var wave = TestHelpers.createWaveform({
-            container: '#test',
-            backend: 'MediaElement'
-        });
-        expect(function() {
-            wave[0].load({
-                foo: 'bar'
-            });
-        }).toThrow(new Error('media parameter is not a valid media element'));
-    });
+    sharedErrorTests('MediaElement');
 });
